Refactor base project structure
