+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".CartActivity"/>
+        <activity android:name=".ui.CartActivity"/>
 
-        <activity android:name=".OnboardingActivity"
+        <activity android:name=".ui.OnboardingActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -20,7 +20,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
             </activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
        
         </activity>
     </application>
