+++ app/src/main/java/com/team19/animation/adapters/CartAdapter.java
@@ -1,4 +1,4 @@
-package com.team19.animation;
+package com.team19.animation.adapters;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -6,12 +6,14 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.AlphaAnimation;
 import android.view.animation.Animation;
 import android.view.animation.ScaleAnimation;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.team19.animation.R;
+import com.team19.animation.models.FoodBean;
 
 import java.util.ArrayList;
