+++ app/src/main/res/layout/activity_cart.xml
@@ -7,7 +7,7 @@
     android:background="#EBE7E7"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CartActivity">
+    tools:context=".ui.CartActivity">
 
 
     <android.support.v7.widget.RecyclerView
