+++ app/src/main/res/layout/activity_onboarding.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context=".OnboardingActivity">
+    tools:context=".ui.OnboardingActivity">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewPager_onboarding"
